requirements.txt
asgiref==3.3.1
Django==3.1.4
django-crispy-forms==1.10.0
django-leaflet==0.27.1
psycopg2-binary==2.8.6
python-google-places==1.4.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
uWSGI==2.0.19.1
